Validando
Adicionando a validação para impedir que usuários sem permissão não entrem nas paginas que necessitam de acesso somente de admin ou usuário.

+++ PHP/DeletarLivros.php
@@ -2,6 +2,11 @@
 require_once 'conexao.php';
 session_start();
 
+if (!isset($_SESSION['id']) || $_SESSION['is_admin'] != 1) {
+    header("Location: ../HTML/login.php");
+    exit();
+}
+
 if($_SERVER['REQUEST_METHOD'] == 'POST'){
     $id = $_POST['id'];
     $sqlDeletarLivro = "DELETE FROM livros WHERE id = :id";
